return value as a float

+++ entity/order_details.go
@@ -22,8 +22,8 @@ type OrderDetails struct {
 } // start from event_transaction
 
 type AdditionalPaymentInfo struct {
-	Name         string `json:"name"`
-	IsTax        bool   `json:"is_tax"`
-	IsPercentage bool   `json:"is_percentage"`
-	Value        int    `json:"value"`
+	Name         string  `json:"name"`
+	IsTax        bool    `json:"is_tax"`
+	IsPercentage bool    `json:"is_percentage"`
+	Value        float64 `json:"value"`
 }
